Implemented Image method via TranVu algorithm. Finished Cython script generating C file when running setup.py. Added two test cases: 1. Comparing RIRs calculated by matlab and python for 2 sensors and 2 sources randomly placed within room dimensions. 2. Comparing timedelay calculated via distance and sound velocity with timedelay read from RIR index of maximum value.

+++ nt/reverb/reverb_utils.py
@@ -88,12 +88,12 @@ def generate_RIR(roomDimension, sourcePositions, sensorPositions, samplingRate,
     # todo: Unterscheide zwischen allen Algorithmen
     # TranVU method
     noiseFloor = -60
-    rir = tranVuRIR.calc(np.asarray(roomDimension,dtype=np.float32),
-                   np.asarray(sourcePositions,dtype=np.float32),
-                   np.asarray(sensorPositions,dtype=np.float32),
+    rir = tranVuRIR.calc(np.asarray(roomDimension,dtype=np.float64),
+                   np.asarray(sourcePositions,dtype=np.float64),
+                   np.asarray(sensorPositions,dtype=np.float64),
                    samplingRate,
                    filterLength,soundDecayTime*1000,noiseFloor,
-                   np.asarray(sensorOrientations, dtype=np.float32 ),
+                   np.asarray(sensorOrientations, dtype=np.float64 ),
                    alpha,soundvelocity)
     return rir
